Fixed classcastexception when casting a blank node to an URI

+++ src/main/java/at/ac/univie/mminf/qskos4j/util/RandomSubSet.java
@@ -1,12 +1,6 @@
 package at.ac.univie.mminf.qskos4j.util;
 
-import java.util.AbstractSet;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Random;
+import java.util.*;
 
 public class RandomSubSet<T> extends AbstractSet<T>{
 
@@ -17,7 +11,7 @@ public class RandomSubSet<T> extends AbstractSet<T>{
 	public RandomSubSet(Collection<T> origSet, float randomSubsetSize_percent) 
 	{
 		origElements = new ArrayList<T>(origSet);	
-		subsetElementCount = (int) ((float) (origSet.size() * randomSubsetSize_percent) / (float) 100);
+		subsetElementCount = (int) ((origSet.size() * randomSubsetSize_percent) / (float) 100);
 		subset = new HashSet<T>(subsetElementCount);
 		
 		pickRandomElements();
